+++ ABS.hpp
@@ -70,7 +70,7 @@ private:
         this->array_ = new T[capacity_];
         T* curr = other.getData(); // points to curr element to copy into new array
         T* temp = array_; // points to curr element of the new array to copy into 
-        for (int i = 0; i < curr_size_; i++)
+        for (size_t i = 0; i < curr_size_; i++)
         {
             *temp = *curr;
             curr++;
@@ -171,7 +171,7 @@ private:
         {
             capacity_ *= scale_factor_;
             T* temp = new T[capacity_];
-            for (int i = 0; i < curr_size_ - 1; i++)
+            for (size_t i = 0; i < curr_size_ - 1; i++)
             {
                 *(temp + i) = *(array_ + i);
             }
